+++ src/xian/services/stamp_calculator.py
